DCKR: fix flapping tests due to Excel metadata
Before this change, the test suite would *sometimes* tell that one or
more tests failed, seemingly at random, and not for every run.
This was due to the Excel file included in the Expectation. The tests
where asserting that Expectations were equals, including the Excel files,
but the `created` and `modified` properties could differ by 1 second
between 2 Excel files created in the same test.

Now we don't compare the Excel files from the Expectations. Ideally we
could have still checked the Excel file content and discard its metadata,
but that would have required either:
- a new library to read Excel files.
- changing our API to have an explicit datetime when generating Excel
  files.
- setting the `created` and `modified` properties to a non accurate
  value.